model op.student.fees.details migrated from openeducat_admission to openeducat_fees module
