+++ openeducat_admission/__manifest__.py
@@ -34,7 +34,6 @@
         'views/admission_register_view.xml',
         'views/admission_view.xml',
         'views/admission_sequence.xml',
-        'views/student_view.xml',
         'report/report_menu.xml',
         'report/report_admission_analysis.xml',
         'wizard/admission_analysis_wizard_view.xml',
