+++ openeducat_admission/models/__init__.py
@@ -21,4 +21,3 @@
 
 from . import admission
 from . import admission_register
-from . import student
