+++ openeducat_admission/security/ir.model.access.csv
@@ -4,6 +4,3 @@ access_op_admission_register_back_office_admin,name_op_admission_register_back_o
 access_op_admission_student,name_op_admission_student,model_op_admission,openeducat_core.group_op_student,1,0,0,0
 access_op_admission_back_office,name_op_admission_back_office,model_op_admission,openeducat_core.group_op_back_office,1,1,1,0
 access_op_admission_back_office_admin,name_op_admission_back_office_admin,model_op_admission,openeducat_core.group_op_back_office_admin,1,1,1,1
-access_op_student_fees_details_student,name_op_student_fees_details_student,model_op_student_fees_details,openeducat_core.group_op_student,1,0,0,0
-access_op_student_fees_details_back_office,name_op_student_fees_details_back_office,model_op_student_fees_details,openeducat_core.group_op_back_office,1,1,1,0
-access_op_student_fees_details_back_office_admin,name_op_student_fees_details_back_office_admin,model_op_student_fees_details,openeducat_core.group_op_back_office_admin,1,1,1,1
