--- openeducat_admission/views/student_view.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record id="view_op_student_inherit_admission_form"
-            model="ir.ui.view">
-            <field name="name">op.student.form</field>
-            <field name="model">op.student</field>
-            <field name="type">form</field>
-            <field name="inherit_id" ref="openeducat_core.view_op_student_form" />
-            <field name="arch" type="xml">
-                <xpath expr="//notebook" position="inside">
-                    <page string="Fees Collection Details">
-                        <field name="fees_detail_ids" readonly="1">
-                            <tree string="Fees Collection Details"
-                                editable="top">
-                                <field name="fees_line_id"
-                                    invisible="1" />
-                                <field name="date" />
-                                <field name="amount" />
-                                <field name="product_id"
-                                    invisible="1" />
-                                <field name="invoice_id"
-                                    invisible="1" />
-                                <field name="state" />
-                                <field name="invoice_state" />
-                                <button string="Create Invoice"
-                                    states="draft" type="object" name="get_invoice"
-                                    groups="openeducat_core.group_op_back_office" />
-                                <button string="View Invoice"
-                                    groups="openeducat_core.group_op_back_office"
-                                    states="invoice" type="object"
-                                    name="action_get_invoice" />
-                            </tree>
-                        </field>
-                    </page>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</openerp>
