+++ openeducat_fees/models/course.py
@@ -26,3 +26,6 @@ class OpCourse(models.Model):
     _inherit = 'op.course'
 
     fees_term_id = fields.Many2one('op.fees.terms', 'Fees Term')
+    
+    product_id = fields.Many2one('product.product', 'Product', domain=[('type', '=', 'service')])
+
