+++ openeducat_fees/security/ir.model.access.csv
@@ -3,3 +3,6 @@ access_op_fees_terms_backoffice,name_op_fees_terms_backoffice,model_op_fees_term
 access_op_fees_terms_backoffice_admin,name_op_fees_terms_backoffice_admin,model_op_fees_terms,openeducat_core.group_op_back_office_admin,1,1,1,1
 access_op_fees_terms_line_backoffice,name_op_fees_terms_line_backoffice,model_op_fees_terms_line,openeducat_core.group_op_back_office,1,1,1,0
 access_op_fees_terms_line_backoffice_admin,name_op_fees_terms_line_backoffice_admin,model_op_fees_terms_line,openeducat_core.group_op_back_office_admin,1,1,1,1
+access_op_student_fees_details_student,name_op_student_fees_details_student,model_op_student_fees_details,openeducat_core.group_op_student,1,0,0,0
+access_op_student_fees_details_back_office,name_op_student_fees_details_back_office,model_op_student_fees_details,openeducat_core.group_op_back_office,1,1,1,0
+access_op_student_fees_details_back_office_admin,name_op_student_fees_details_back_office_admin,model_op_student_fees_details,openeducat_core.group_op_back_office_admin,1,1,1,1
