+++ openeducat_fees/views/course_view.xml
@@ -20,6 +20,7 @@
             <field name="arch" type="xml">
                 <field name="evaluation_type" position="after">
                     <field name="fees_term_id" required="1" />
+                    <field name="product_id" required="1" />
                 </field>
             </field>
         </record>
